Add files via upload
This is the python script that generates diet schedule.

It selects at random 6 diets from a list of "regular" diets, and an additional diet from a list of "special" diets. (These are diets that include 'treat' meals such as a cheeseburger, pizza, etc...). All diets in both lists come from a collection of diets that was given to me by a licensed doctor, and so are customized for my own personal use. A diet is a dictionary with three key, value pairs: in which each key is a meal-time ("breakfast", "lunch", or "dinner"), and each value is a meal for that time.

The special diet is assigned to Saturday, and all other days of the week get a regular diet. The output is a schedule for all days of the week starting on Monday and ending on Sunday.